Egress Firewall Rule applied for Fusion storage namespace

* make fmt runned

* Lint error reseolved

* golangci-lint check resolved

* removed comments in egressfirewall_offering.go
