+++ api/v1alpha1/groupversion_info.go
@@ -15,8 +15,8 @@ limitations under the License.
 */
 
 // Package v1alpha1 contains API Schema definitions for the misf v1alpha1 API group
-//+kubebuilder:object:generate=true
-//+groupName=misf.ibm.com
+// +kubebuilder:object:generate=true
+// +groupName=misf.ibm.com
 package v1alpha1
 
 import (
